move 'and' tests to separate file
